+++ src/one_app_to_rule_them_all.py
@@ -16,6 +16,7 @@ from PyQt6.QtWidgets import (
     QVBoxLayout,
     QGridLayout,
     QFormLayout,
+    QFrame
 )
 
 
@@ -155,6 +156,7 @@ class LiuWindow(QMainWindow):
         graph_scroll = QScrollArea()
         graph_scroll.setWidgetResizable(True)
         graph_scroll.setFixedSize(SCROLL_WIDTH, SCROLL_HEIGHT)
+        graph_scroll.setFrameShape(QFrame.Shape.NoFrame)
         graph_scroll.setWidget(graph_values_widget)
 
         self.graph_quarter_layout = QVBoxLayout()
@@ -180,9 +182,9 @@ class LiuWindow(QMainWindow):
         print("[Start] create_tasks")
         tasks_list = []
         for el in range(tasks_number):
-            execution_time = random.randint(min_time, max_time)
+            execution_time = random.randint(min_time, int(max_time/3))
             release_time = random.randint(min_time, max_time)
-            deadline = release_time + execution_time - random.randint(min_time, max_time)
+            deadline = release_time + execution_time + int(3 / 2 * random.randint(min_time, int(max_time/3)))
 
             tasks_list.append(Task(p=execution_time, r=release_time, d=deadline))
         print("[End]   create_tasks")