+++ src/one_app_to_rule_them_all.py
@@ -3,7 +3,8 @@ import random
 from functools import partial
 
 from src.configuration.config import DEFAULT_TASK_VALUES
-from src.configuration.config import WINDOW_HEIGHT, WINDOW_WIDTH, SCROLL_HEIGHT, SCROLL_WIDTH, COLOR_PALETTE, EMPTY_COLOR, BUTTON_WIDTH, BUTTON_HEIGHT, LINE_EDIT_SIZE, TASK_LABEL_WIDTH, APP_BG_COLOR, BUTTON_COLOR
+from src.configuration.config import WINDOW_HEIGHT, WINDOW_WIDTH, SCROLL_HEIGHT, SCROLL_WIDTH, COLOR_PALETTE, \
+    EMPTY_COLOR, BUTTON_WIDTH, BUTTON_HEIGHT, LINE_EDIT_SIZE, TASK_LABEL_WIDTH, APP_BG_COLOR, BUTTON_COLOR
 from PyQt6.QtCore import Qt
 from PyQt6.QtWidgets import (
     QApplication,
@@ -99,21 +100,30 @@ class LiuWindow(QMainWindow):
         self.button_default = QPushButton("Create default tasks")
         self.button_run_alg = QPushButton("Order tasks")
 
+        # invalid params label
+        self.invalid_params_label = QLabel("")
+        self.invalid_params_label.setStyleSheet("font-size:15px; color:red")
+
         self.button_user.setFixedSize(BUTTON_WIDTH, BUTTON_HEIGHT)
         self.button_default.setFixedSize(BUTTON_WIDTH, BUTTON_HEIGHT)
         self.button_run_alg.setFixedSize(BUTTON_WIDTH, BUTTON_HEIGHT)
-        self.button_user.setStyleSheet("font-size:15px; border-radius : 5; background-color: #cfeafb; color:black; border-bottom: 1px solid grey; border-right:1px solid grey;")
-        self.button_default.setStyleSheet("font-size:15px; border-radius : 5; background-color: #cfeafb; color:black; border-bottom: 1px solid grey; border-right:1px solid grey;")
-        self.button_run_alg.setStyleSheet("font-size:15px; border-radius : 5; background-color: #cfeafb; color:black; border-bottom: 1px solid grey; border-right:1px solid grey;")
+        self.button_user.setStyleSheet(
+            "font-size:15px; border-radius : 5; background-color: #cfeafb; color:black; border-bottom: 1px solid grey; border-right:1px solid grey;")
+        self.button_default.setStyleSheet(
+            "font-size:15px; border-radius : 5; background-color: #cfeafb; color:black; border-bottom: 1px solid grey; border-right:1px solid grey;")
+        self.button_run_alg.setStyleSheet(
+            "font-size:15px; border-radius : 5; background-color: #cfeafb; color:black; border-bottom: 1px solid grey; border-right:1px solid grey;")
 
         # add user input layout to user quarter layout
         self.user_quarter_layout = QVBoxLayout()
         self.user_quarter_layout.addWidget(QLabel("<h1>Algorithm Liu</h1>"), alignment=Qt.AlignmentFlag.AlignCenter)
-        self.user_quarter_layout.addWidget(QLabel("<h2>1 | r<sub>i</sub>, prm | L<sub>max</sub></h2>"), alignment=Qt.AlignmentFlag.AlignCenter)
+        self.user_quarter_layout.addWidget(QLabel("<h2>1 | r<sub>i</sub>, prm | L<sub>max</sub></h2>"),
+                                           alignment=Qt.AlignmentFlag.AlignCenter)
         self.user_quarter_layout.addWidget(user_input_widget, alignment=Qt.AlignmentFlag.AlignCenter)
         self.user_quarter_layout.addWidget(self.button_user, alignment=Qt.AlignmentFlag.AlignCenter)
         self.user_quarter_layout.addWidget(self.button_default, alignment=Qt.AlignmentFlag.AlignCenter)
         self.user_quarter_layout.addWidget(self.button_run_alg, alignment=Qt.AlignmentFlag.AlignCenter)
+        self.user_quarter_layout.addWidget(self.invalid_params_label, alignment=Qt.AlignmentFlag.AlignCenter)
 
         # add user's layout to user widget
         self.general_layout.addLayout(self.user_quarter_layout, 0, 0, alignment=Qt.AlignmentFlag.AlignCenter)
@@ -165,15 +175,37 @@ class LiuWindow(QMainWindow):
 
         self.general_layout.addLayout(self.graph_quarter_layout, 1, 1)
 
+    def _clear_user_values(self):
+        print("[Start] _clear_user_values")
+        self.user_min_time.setText("")
+        self.user_max_time.setText("")
+        self.user_tasks_number.setText("")
+        self.user_min_time.setFocus()
+        print("[End]   _clear_user_values")
+
     def get_user_values(self) -> dict:
         print("[Start] get_user_values")
         user_values = {
-            "min_time": int(self.user_min_time.text()),
-            "max_time": int(self.user_max_time.text()),
-            "tasks_number": int(self.user_tasks_number.text())
+            "min_time": self.user_min_time.text(),
+            "max_time": self.user_max_time.text(),
+            "tasks_number": self.user_tasks_number.text()
         }
         print(f"[End] get_user_values, returning: {user_values}")
         return user_values
+
+    def _validate_user_values(self, user_values):
+        if user_values['min_time'].isnumeric() and \
+                user_values['max_time'].isnumeric() and \
+                user_values['tasks_number'].isnumeric() and \
+                int(user_values['tasks_number']) <= 10:
+            print("Validation - returning TRUE")
+            return True
+        else:
+            print("Validation - returning Flase")
+            return False
+
+    def _convert_values_to_int(self, values):
+        return {key: int(value) for key, value in values.items()}
 
     def create_tasks(self, **kwargs):
         min_time = kwargs["min_time"]
@@ -182,9 +214,9 @@ class LiuWindow(QMainWindow):
         print("[Start] create_tasks")
         tasks_list = []
         for el in range(tasks_number):
-            execution_time = random.randint(min_time, int(max_time/3))
+            execution_time = random.randint(min_time, int(max_time / 3))
             release_time = random.randint(min_time, max_time)
-            deadline = release_time + execution_time + int(3 / 2 * random.randint(min_time, int(max_time/3)))
+            deadline = release_time + execution_time + int(3 / 2 * random.randint(min_time, int(max_time / 3)))
 
             tasks_list.append(Task(p=execution_time, r=release_time, d=deadline))
         print("[End]   create_tasks")
@@ -227,25 +259,32 @@ class LiuWindow(QMainWindow):
     def generate_input(self, creation_method):
         print("[Start] generate_input")
         if creation_method == self.create_tasks:
-            print("Creation method is: user method")
             user_values = self.get_user_values()
-            self.tasks_list = creation_method(
-                min_time=user_values.get("min_time"),
-                max_time=user_values.get("max_time"),
-                tasks_number=user_values.get("tasks_number")
-            )
+            validation_pass = self._validate_user_values(user_values)
+            if validation_pass:
+                user_values = self._convert_values_to_int(user_values)
+                self.invalid_params_label.setText("")
+                self.tasks_list = creation_method(
+                    min_time=user_values.get("min_time"),
+                    max_time=user_values.get("max_time"),
+                    tasks_number=user_values.get("tasks_number")
+                )
+            else:
+                self._clear_user_values()
+                self.invalid_params_label.setText("Invalid parameters! Try again.")
         elif creation_method == self.create_tasks_default:
-            print("Creation method is: default method")
             self.tasks_list = creation_method(values=DEFAULT_TASK_VALUES)
         else:
-            raise Exception(f"Given tasks creation method:\n\t{creation_method}\nis not a user creation method:\n\t{self.create_tasks}\nand not a default creation method:\n\t{self.create_tasks_default}")
+            raise Exception(
+                f"Given tasks creation method:\n\t{creation_method}\nis not a user creation method:\n\t{self.create_tasks}\nand not a default creation method:\n\t{self.create_tasks_default}")
 
-        tasks_print_table = []
-        for task in self.tasks_list:
-            task_row = [task, task.execution_time, task.release_time, task.deadline]
-            tasks_print_table.append(task_row)
-        # print data
-        self.fill_input_grid(tasks_print_table)
+        # TODO: move below to "fill_input_grid" method
+        if self.tasks_list:
+            tasks_print_table = []
+            for task in self.tasks_list:
+                task_row = [task, task.execution_time, task.release_time, task.deadline]
+                tasks_print_table.append(task_row)
+            self.fill_input_grid(tasks_print_table)
         print("[End]   generate_input")
 
     def perform_alg(self, tasks):
@@ -333,7 +372,8 @@ class LiuWindow(QMainWindow):
             time_label.setAlignment(Qt.AlignmentFlag.AlignLeft)
             self.graph_values_layout.addWidget(time_label, 1, idx)
 
-        self.graph_quarter_layout.addWidget(QLabel(f"<h2>L<sub>max</sub>= {lmax}</h2>"), alignment=Qt.AlignmentFlag.AlignCenter)
+        self.graph_quarter_layout.addWidget(QLabel(f"<h2>L<sub>max</sub>= {lmax}</h2>"),
+                                            alignment=Qt.AlignmentFlag.AlignCenter)
         print("[End]   fill_graph_layout")
 
     def order_tasks(self):