+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 - App's window size changed to 1600 to handle max (=10) amount of tasks.
 - Remove graph scroller border.
+- Text size for user input lines, size of input lines.
 
 ## Added